Camera movement along axis X done

+++ freeglut project 3D/sources/PV3D.cpp
@@ -12,7 +12,7 @@ PV3D::PV3D(void)
 	this->pv = 1;
 }
 
-PV3D::PV3D(GLfloat x, GLfloat y, GLfloat z, int pv)
+PV3D::PV3D(GLdouble x, GLdouble y, GLdouble z, int pv)
 {
 	this->x = x;
 	this->y = y;
@@ -20,14 +20,14 @@ PV3D::PV3D(GLfloat x, GLfloat y, GLfloat z, int pv)
 	this->pv = pv;
 }
 
-GLfloat PV3D::module(PV3D* v){
+GLdouble PV3D::module(PV3D* v){
 	return sqrt(pow(v->getX(), 2) + pow(v->getY(), 2) + pow(v->getZ(), 2));
 }
 
 PV3D* PV3D::normaliza(){
 	// Same direction but module = 1
 	PV3D* newV = new PV3D();
-	GLfloat moduleVector = module(this);
+	GLdouble moduleVector = module(this);
 
 	newV->setX((1 / moduleVector) * this->getX());
 	newV->setY((1 / moduleVector) * this->getY());
@@ -37,34 +37,34 @@ PV3D* PV3D::normaliza(){
 	return newV;
 }
 
-GLfloat PV3D::productoEscalar(PV3D* v){
+GLdouble PV3D::productoEscalar(PV3D* v){
 	return (this->getX() * v->getX()) + (this->getY() * v->getY() + this->getZ() * v->getZ());
 }
 
 PV3D* PV3D::productoVectorial(PV3D* v){
-	GLfloat x = this->getY()*v->getZ() - this->getZ()*v->getY();
-	GLfloat y = this->getZ()*v->getX() - this->getX()*v->getZ();
-	GLfloat z = this->getX()*v->getY() - this->getY()*v->getX();
+	GLdouble x = this->getY()*v->getZ() - this->getZ()*v->getY();
+	GLdouble y = this->getZ()*v->getX() - this->getX()*v->getZ();
+	GLdouble z = this->getX()*v->getY() - this->getY()*v->getX();
 
 	return new PV3D(x,y,z,1);
 }
 
 PV3D* PV3D::calculaValorFormula(float t){
-	GLfloat x = 3 * cos(t);
-	GLfloat y = 2 * cos(1.5 * t);
-	GLfloat z = 3 * sin(t);
+	GLdouble x = 3 * cos(t);
+	GLdouble y = 2 * cos(1.5 * t);
+	GLdouble z = 3 * sin(t);
 	return new PV3D(x,y,z,1);
 }
 PV3D* PV3D::calculaValorPrimDerivada(float t){
-	GLfloat x = -3 * sin(t);
-	GLfloat y = -3 * sin(1.5 * t);
-	GLfloat z = 3 * cos(t);
+	GLdouble x = -3 * sin(t);
+	GLdouble y = -3 * sin(1.5 * t);
+	GLdouble z = 3 * cos(t);
 	return new PV3D(x,y,z,1);
 }
 PV3D* PV3D::calculaValorSegDerivada(float t){
-	GLfloat x = -3 * cos(t);
-	GLfloat y = -4.5 * cos(1.5 * t);
-	GLfloat z = -3 * sin(t);
+	GLdouble x = -3 * cos(t);
+	GLdouble y = -4.5 * cos(1.5 * t);
+	GLdouble z = -3 * sin(t);
 	return new PV3D(x,y,z,1);
 }
